agrisense-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.30.0
redis>=5.0.1
pinecone>=5.0.0
google-generativeai>=0.4.0
elevenlabs>=1.0.0
pillow>=10.2.0
opencv-python>=4.9.0
sentence-transformers>=2.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
